inference/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
transformers>=4.49.0
torch>=2.2.0
accelerate>=0.27.0
fasttext-wheel>=0.9.2
pydantic>=2.5.0
huggingface_hub>=0.20.0
python-dotenv>=1.0.0

# Voice translation dependencies
faster-whisper>=1.0.0
websockets>=12.0
soundfile>=0.12.0
librosa>=0.10.0
scipy>=1.11.0
numpy>=1.24.0

# Optional: Speaker diarization (requires HuggingFace token)
# pyannote.audio>=3.0

# Optional: TTS (CosyVoice requires manual install)
# cosyvoice @ git+https://github.com/FunAudioLLM/CosyVoice.git
edge-tts>=6.1.0
